Updated dd_edge::getCardinality() to return a double.
git-svn-id: svn://svn.code.sf.net/p/meddly/code/trunk@38 14966d26-f10a-4722-9547-d5dcc201c352

+++ examples/kan1_batch.cc
@@ -326,8 +326,8 @@ int main(int argc, char* argv[])
   getIndexSet(reachableStates, indexSet);
   int* element = (int *) malloc(N * sizeof(int));
 
-  int cardinality = indexSet.getCardinality();
-  for (int index = 0; index < cardinality; index++)
+  double cardinality = indexSet.getCardinality();
+  for (int index = 0; index < int(cardinality); index++)
   {
     assert(forest::SUCCESS == evplusmdd->getElement(indexSet, index, element));
     printf("Element at index %d: [ ", index);
